moved github deep investigation into own folder
